+++ firmware/controllers/algo/fuel/fuel_computer.h
@@ -10,20 +10,20 @@ class ValueProvider3D;
 #include "fuel_computer_generated.h"
 
 struct IFuelComputer : public fuel_computer_s {
-	virtual mass_t getCycleFuel(mass_t airmass, int rpm, float load) = 0;
-	temperature_t getTCharge(int rpm, float tps);
+	virtual mass_t getCycleFuel(mass_t airmass, float rpm, float load) = 0;
+	temperature_t getTCharge(float rpm, float tps);
 	float getLoadOverride(float defaultLoad, load_override_e overrideMode) const;
 private:
-	float getTChargeCoefficient(int rpm, float tps);
+	float getTChargeCoefficient(float rpm, float tps);
 };
 
 // This contains the math of the fuel model, but doesn't actually read any configuration
 class FuelComputerBase : public IFuelComputer {
 public:
-	mass_t getCycleFuel(mass_t airmass, int rpm, float load) override;
+	mass_t getCycleFuel(mass_t airmass, float rpm, float load) override;
 
 	virtual float getStoichiometricRatio() const = 0;
-	virtual float getTargetLambda(int rpm, float load) const = 0;
+	virtual float getTargetLambda(float rpm, float load) const = 0;
 	virtual float getTargetLambdaLoadAxis(float defaultLoad) const = 0;
 };
 
@@ -31,7 +31,7 @@ public:
 class FuelComputer final : public FuelComputerBase {
 public:
 	float getStoichiometricRatio() const override;
-	float getTargetLambda(int rpm, float load) const override;
+	float getTargetLambda(float rpm, float load) const override;
 	float getTargetLambdaLoadAxis(float defaultLoad) const override;
 };
